requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.2.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.2.0
colorama==0.4.6
comm==0.2.2
cookiecutter==2.6.0
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
fastapi==0.115.12
fastjsonschema==2.20.0
flet==0.28.2
flet-cli==0.28.2
flet-desktop==0.28.2
flet-web==0.28.2
fqdn==1.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
oauthlib==3.2.2
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pdfkit==1.0.0
pillow==11.2.1
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.2.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
python-slugify==8.0.4
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==7.4.2
referencing==0.35.1
repath==0.9.0
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.21.0
rsa==4.9.1
Send2Trash==1.8.3
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.2
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.0
typing_extensions==4.13.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.2
watchdog==4.0.2
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
